financial_bookkeeping/requirements.txt
Flask==3.1.0
python_dateutil==2.9.0.post0
